Туча доработок
- класс `PayingAgent`, покрытый тестами
- новые константы для тегов ФФД 1.05 `Ffd105Tags`
- `Entity::jsonSerialize()` object -> array (again)
- `TooManyException::$max` int -> float
- тесты по psr-4, потому что почему бы и нет
- некоторые провайдеры вынесены в `BasicTestCase`
- улучшен тест покупателя

+++ src/Exceptions/TooLongCallbackUrlException.php
@@ -19,5 +19,5 @@ use AtolOnline\Constants\Constraints;
 class TooLongCallbackUrlException extends TooLongException
 {
     protected $message = 'Слишком длинный адрес колбека';
-    protected int $max = Constraints::MAX_LENGTH_CALLBACK_URL;
+    protected float $max = Constraints::MAX_LENGTH_CALLBACK_URL;
 }
